requirements.txt
africastalking==1.2.9
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.6.10
cryptography==44.0.0
Django==5.1.4
djangorestframework==3.15.2
ecdsa==0.19.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
